Ctrl+C後に送信を行うとサーバ側で文字化けが発生する

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: smihata <smihata@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/08 17:00:52 by smihata           #+#    #+#             */
-/*   Updated: 2023/03/09 11:53:36 by smihata          ###   ########.fr       */
+/*   Updated: 2023/08/19 16:07:49 by smihata          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,7 +58,7 @@ void	ft_striteri(char *s, void (*f)(unsigned int, char *));
 void	ft_putchar_fd(char c, int fd);
 void	ft_putstr_fd(char *s, int fd);
 void	ft_putendl_fd(char *s, int fd);
-void	ft_putnbr_fd(int n, int fd);
+void	ft_putnbr_fd(long long n, int fd);
 
 // Bonus
 t_list	*ft_lstnew(void *content);
